Added STEP version of files
These files can't be as easily modified as the Solidworks files, but they can be directly used for fabrication with JLCCNC.com. The putter would consist of 1x FLAB body, 1 x Face insert, and 1 x of whatever shaft block lie angle would work best for your setup. The soft jaws are for milling the face to 3 degrees of loft and should be 3D printed out of some sort of plastic.